logging drink service

+++ shop-app/src/main/java/com/spring/shopapp/service/DrinkService.java
@@ -2,6 +2,8 @@ package com.spring.shopapp.service;
 
 import com.spring.shopapp.entity.Drinks;
 import com.spring.shopapp.repository.DrinksRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -16,26 +18,46 @@ import java.util.Optional;
 @Service
 public class DrinkService {
 
+    private static Logger LOGGER = LoggerFactory.getLogger(DrinkService.class);
+
     @Autowired
     DrinksRepository drinksRepository;
 
-    public Drinks getDrinkById(Long id) {
-        return drinksRepository.getById(id);
+    public ResponseEntity<Drinks> getDrinkById(Long id) {
+        LOGGER.debug("Searching drink with id: " + id);
+
+        Optional<Drinks> drink = drinksRepository.findById(id);
+
+        if(drink.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            return new ResponseEntity<>(drink.get(), HttpStatus.NOT_FOUND);
+        }
+        LOGGER.info("Drink with id: " + id + " was found successfully.");
+        return new ResponseEntity<>(drink.get(), HttpStatus.OK);
     }
 
     public ResponseEntity<List<Drinks>> getDrinkByName(String name) {
+        LOGGER.debug("Searching drink with name: " + name);
+
         List<Drinks> drinkName = drinksRepository.findDrinkByName(name);
-        return drinkName.isEmpty() ? new ResponseEntity<>(drinkName, HttpStatus.NOT_FOUND) :
-                new ResponseEntity<>(drinkName, HttpStatus.OK);
+
+        if(drinkName.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            new ResponseEntity<>(drinkName, HttpStatus.NOT_FOUND);
+        }
+        LOGGER.info("Drink with name: " + name + " was found successfully.");
+        return new ResponseEntity<>(drinkName, HttpStatus.OK);
     }
 
     public ResponseEntity<List<Drinks>> findAllDrinks() {
+        LOGGER.debug("Searching all drinks...");
         List<Drinks> drinks = new ArrayList<>(drinksRepository.findAll());
         return new ResponseEntity<>(drinks, HttpStatus.OK);
     }
 
     public ResponseEntity<Drinks> addDrink(Drinks drink) {
         drinksRepository.save(drink);
+        LOGGER.info(drink + " was saved successfully.");
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
@@ -43,36 +65,60 @@ public class DrinkService {
         Optional<Drinks> existingDrink = drinksRepository.findById(id);
 
         if(existingDrink.isEmpty()) {
+            LOGGER.error("Drink not found.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         BeanUtils.copyProperties(drink, existingDrink.get(), "id");
         drinksRepository.save(existingDrink.get());
+        LOGGER.info(drink + " was updated successfully.");
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     public ResponseEntity<Drinks> findDrinkByPrice(BigDecimal price) {
+        LOGGER.debug("Searching drink with price: " + price);
         Optional<Drinks> drinkPrice = drinksRepository.findByPrice(price);
 
-        return drinkPrice.isPresent() ? new ResponseEntity<>(drinkPrice.get(), HttpStatus.OK) :
-                new ResponseEntity<>(drinkPrice.get(), HttpStatus.NOT_FOUND);
+        if(drinkPrice.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            return new ResponseEntity<>(drinkPrice.get(), HttpStatus.NOT_FOUND);
+        }
+        LOGGER.info("Drink with price: " + price + " was found successfully.");
+        return  new ResponseEntity<>(drinkPrice.get(), HttpStatus.OK);
     }
 
     public ResponseEntity<List<Drinks>> findDrinkByManufacturerName(String manufacturerName) {
+        LOGGER.debug("Searching drink with manufacturer name: " + manufacturerName);
         List<Drinks> drink = drinksRepository.findByManufacturerName(manufacturerName);
 
-        return drink.isEmpty() ? new ResponseEntity<>(drink, HttpStatus.NOT_FOUND) :
-                new ResponseEntity<>(drink, HttpStatus.OK);
+        if(drink.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            return new ResponseEntity<>(drink, HttpStatus.NOT_FOUND);
+        }
+        LOGGER.info("Drink with manufacturerName: " + manufacturerName + " was found successfully.");
+        return new ResponseEntity<>(drink, HttpStatus.OK);
     }
 
     public ResponseEntity<Drinks> findDrinksByAlcoholContent(double alcoholContent) {
+        LOGGER.debug("Searching drink with alcohol content: " + alcoholContent);
         Optional<Drinks> drink = drinksRepository.findByAlcoholContent(alcoholContent);
 
-        return drink.isPresent() ? new ResponseEntity<>(drink.get(), HttpStatus.OK) :
-                new ResponseEntity<>(drink.get(), HttpStatus.NOT_FOUND);
+        if(drink.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            return new ResponseEntity<>(drink.get(), HttpStatus.NOT_FOUND);
+        }
+        LOGGER.info("Drink with alcoholContent: " + alcoholContent + " was found successfully.");
+        return new ResponseEntity<>(drink.get(), HttpStatus.OK);
     }
 
     public ResponseEntity<Drinks> deleteDrinkById(Long id) {
+        Optional<Drinks> drink = drinksRepository.findById(id);
+
+        if(drink.isEmpty()) {
+            LOGGER.error("Drink not found.");
+            return new ResponseEntity<>(drink.get(), HttpStatus.NOT_FOUND);
+        }
         drinksRepository.deleteById(id);
+        LOGGER.info("Drink was deleted successfully.");
         return new ResponseEntity<>(HttpStatus.OK);
     }
 }
